fix(combinator): Pull in multi
