+++ examples/http/parser.rs
@@ -1,4 +1,4 @@
-use winnow::{character::line_ending, multi::many1, token::take_while1, IResult, Parser};
+use winnow::{character::line_ending, combinator::many1, token::take_while1, IResult, Parser};
 
 pub type Stream<'i> = &'i [u8];
