+++ examples/ini/parser_str.rs
@@ -3,9 +3,9 @@ use std::collections::HashMap;
 use winnow::prelude::*;
 use winnow::{
     character::{alphanumeric1 as alphanumeric, space0 as space},
+    combinator::many0,
     combinator::opt,
     combinator::{delimited, terminated},
-    multi::many0,
     token::{take_till0, take_while0, take_while1},
 };
