+++ examples/json/parser_dispatch.rs
@@ -10,8 +10,8 @@ use winnow::{
     combinator::success,
     combinator::{alt, dispatch},
     combinator::{delimited, preceded, separated_pair, terminated},
+    combinator::{fold_many0, separated0},
     error::{ContextError, ParseError},
-    multi::{fold_many0, separated0},
     token::{any, none_of, take, take_while0},
 };
