+++ examples/json_iterator.rs
@@ -5,9 +5,9 @@ use winnow::{
     character::{alphanumeric1 as alphanumeric, escaped, float},
     combinator::alt,
     combinator::cut_err,
+    combinator::separated0,
     combinator::{preceded, separated_pair, terminated},
     error::ParseError,
-    multi::separated0,
     token::one_of,
     token::{tag, take_while0},
     IResult,
