+++ examples/s_expression/parser.rs
@@ -5,10 +5,10 @@
 use winnow::{
     character::{alpha1, digit1, multispace0, multispace1},
     combinator::alt,
+    combinator::many0,
     combinator::{cut_err, opt},
     combinator::{delimited, preceded, terminated},
     error::VerboseError,
-    multi::many0,
     token::one_of,
     IResult, Parser,
 };
